requirements.txt
django==4.1.1
django-bootstrap4==3.0.1
django-filter==2.4.0
django-import-export
django-tables2==2.4.0
gunicorn==19.9.0
pandas
psycopg2-binary==2.8.6
python-dotenv==0.19.2
tablib
whitenoise==5.0.1
